Implement user authentication and connect to backend
Implemented user registration and login functionalities. Added new "Register" component for user signups. Updated "Login" component for verification against backend. Facilitated authenticated requests by storing JWT token in local storage and using it as the authorization header. Configured Axios to use a relative path for API calls and added "proxy" in package.json to direct relative paths to the specified API server. The "proxy" was changed from an external link, to use the internal backend server.

Connected to the backend by creating a Go backend with Postgres database setup in Docker, including a Dockerfile and docker-compose.yml file for containerisation. Organized the Go backend files for ease of reference.

Adjusted the check for whether the user is logged in in App.js and added a check to delete the token from local storage on logout. Made further modifications to other components to integrate the authentication token into different requests. Added Go backend files to .gitignore, as they will be built and run in a Docker container, and are not necessary to include in commits.

This update is preparation for integrating the application with a backend, which provides user authentication and persistent storage.

Additional notable changes include usability and visual enhancements such as improved text alignment in "CaughtBreeds" and added styles for "Register" component. The UI changes are intended to improve the user experience and overall aesthetic of the application.